// src/app/modules/account-module/models/inv-commision-model.model.ts
export class InvCommisionModel {
    ComInvId :number = 0;
    BranchId : number = 0;
    CompanyId : number = 0;
    InvoiceNumber: number = 0;
    InvoiceDate: Date = new Date();
    InvType: number = 0;
    Company: string = "";
    CompanyAddress: string = "";
    Description: string = "";
    TaxId:number=0;
    CompanyCity: string = "";
    TaxableAmount: number = 0;
    CompanyName: string = "";
    CGST:number = 0;
    SGST:number = 0;
    TotalAmt: number = 0;
    CompanyCode: number = 0;
    MasterName: string = ""
    InvTypeName: string = '';
    Action:string="";
    Addedby:string="";
    LastUpdatedOn:string="";
    InvTypeId:number=0;
}

       
       
      
        
       
       
       
       